mypage flask-apispec update
