+++ aimmoPost/tests/views/test_board.py
@@ -119,65 +119,3 @@ class Test_BoardView:
 
         def test_삭제_여부(self, board, trans_api):
             assert len(Board.objects(id=board.id)) == 0
-
-    class Test_Get_List:
-        @pytest.fixture
-        def post(self, login_user, board):
-            return [PostFactory.create(like=random.choice([[], [login_user]]), notice=random.choice([True, False]),
-                                       board=board) for _ in range(20)]
-
-        @pytest.fixture
-        def trans_api(self, client, board, headers, post):
-            return client.get(f"/boards/{str(board.id)}/posts/?page=1&size=10", headers=headers)
-
-        def test_상태코드_200(self, trans_api):
-            assert trans_api.status_code == 200
-
-        def test_데이터_확인(self, trans_api):
-            posts = trans_api.json["post_list"]
-            for i in posts:
-                assert "id" in i
-                assert "writer" in i
-                assert "date" in i
-                assert "title" in i
-                assert "content" in i
-                assert "tag" in i
-                assert "notice" in i
-                assert "num_like" in i
-                assert "num_comment" in i
-
-        def test_공지사항_상단_노출_여부(self, trans_api):
-            posts = trans_api.json["post_list"]
-            for i in range(1, len(posts)):
-                assert (posts[i - 1]["notice"] == True and posts[i]["notice"] == True) or (
-                            posts[i - 1]["notice"] == False and posts[i]["notice"] == False) \
-                       or (posts[i - 1]["notice"] == True and posts[i]["notice"] == False)
-
-        def test_존재_여부(self, trans_api, post):
-            posts = trans_api.json["post_list"]
-            for i in posts:
-                assert next((True for item in post if str(item.id) == i["id"]), False)
-
-        def test_좋아요_누름_여부(self, trans_api, login_user):
-            posts = trans_api.json["post_list"]
-            for i in posts:
-                assert (login_user in Post.objects(id=i["id"]).get().like) == i["is_like"]
-
-        class Test_페이지_인덱스가_음수일_경우:
-            @pytest.fixture
-            def trans_api(self, client, board, headers):
-                return client.get(f"/boards/{str(board.id)}/posts/?page=-1&size=10", headers=headers)
-
-            def test_상태코드_404(self, trans_api):
-                assert trans_api.status_code == 404
-
-        class Test_페이지_인덱스가_매우_클경우:
-            @pytest.fixture
-            def trans_api(self, client, board, headers):
-                return client.get(f"/boards/{str(board.id)}/posts/?page=100&size=10", headers=headers)
-
-            def test_상태코드_404(self, trans_api):
-                assert trans_api.status_code == 200
-
-            def test_빈리스트_반환(self, trans_api):
-                assert trans_api.json["post_list"] == []
